[improvement][chat]llm parser corrector is simplified by sql distribution
