--- chat/core/src/main/java/com/tencent/supersonic/chat/corrector/FieldValueCorrector.java
@@ -1,81 +1,0 @@
-package com.tencent.supersonic.chat.corrector;
-
-import com.tencent.supersonic.chat.api.pojo.SemanticCorrectInfo;
-import com.tencent.supersonic.chat.api.pojo.SchemaElement;
-import com.tencent.supersonic.chat.api.pojo.SchemaValueMap;
-import com.tencent.supersonic.chat.api.pojo.SemanticSchema;
-import com.tencent.supersonic.common.util.ContextUtils;
-import com.tencent.supersonic.common.util.jsqlparser.SqlParserUpdateHelper;
-import com.tencent.supersonic.knowledge.service.SchemaService;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.stream.Collectors;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.logging.log4j.util.Strings;
-import org.springframework.util.CollectionUtils;
-
-@Slf4j
-public class FieldValueCorrector extends BaseSemanticCorrector {
-
-    @Override
-    public void correct(SemanticCorrectInfo semanticCorrectInfo) {
-        SemanticSchema semanticSchema = ContextUtils.getBean(SchemaService.class).getSemanticSchema();
-        Long modelId = semanticCorrectInfo.getParseInfo().getModel().getId();
-        List<SchemaElement> dimensions = semanticSchema.getDimensions().stream()
-                .filter(schemaElement -> modelId.equals(schemaElement.getModel()))
-                .collect(Collectors.toList());
-
-        if (CollectionUtils.isEmpty(dimensions)) {
-            return;
-        }
-
-        Map<String, Map<String, String>> aliasAndBizNameToTechName = getAliasAndBizNameToTechName(dimensions);
-        String preSql = semanticCorrectInfo.getSql();
-        semanticCorrectInfo.setPreSql(preSql);
-        String sql = SqlParserUpdateHelper.replaceValue(preSql, aliasAndBizNameToTechName);
-        semanticCorrectInfo.setSql(sql);
-        return;
-    }
-
-
-    private Map<String, Map<String, String>> getAliasAndBizNameToTechName(List<SchemaElement> dimensions) {
-        if (CollectionUtils.isEmpty(dimensions)) {
-            return new HashMap<>();
-        }
-
-        Map<String, Map<String, String>> result = new HashMap<>();
-
-        for (SchemaElement dimension : dimensions) {
-            if (Objects.isNull(dimension)
-                    || Strings.isEmpty(dimension.getBizName())
-                    || CollectionUtils.isEmpty(dimension.getSchemaValueMaps())) {
-                continue;
-            }
-            String bizName = dimension.getBizName();
-
-            Map<String, String> aliasAndBizNameToTechName = new HashMap<>();
-
-            for (SchemaValueMap valueMap : dimension.getSchemaValueMaps()) {
-                if (Objects.isNull(valueMap) || Strings.isEmpty(valueMap.getTechName())) {
-                    continue;
-                }
-                if (Strings.isNotEmpty(valueMap.getBizName())) {
-                    aliasAndBizNameToTechName.put(valueMap.getBizName(), valueMap.getTechName());
-                }
-                if (!CollectionUtils.isEmpty(valueMap.getAlias())) {
-                    valueMap.getAlias().stream().forEach(alias -> {
-                        if (Strings.isNotEmpty(alias)) {
-                            aliasAndBizNameToTechName.put(alias, valueMap.getTechName());
-                        }
-                    });
-                }
-            }
-            if (!CollectionUtils.isEmpty(aliasAndBizNameToTechName)) {
-                result.put(bizName, aliasAndBizNameToTechName);
-            }
-        }
-        return result;
-    }
-}
