+++ chat/core/src/main/java/com/tencent/supersonic/chat/corrector/GroupByCorrector.java
@@ -1,0 +1,15 @@
+package com.tencent.supersonic.chat.corrector;
+
+import com.tencent.supersonic.chat.api.pojo.SemanticCorrectInfo;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+public class GroupByCorrector extends BaseSemanticCorrector {
+
+    @Override
+    public void correct(SemanticCorrectInfo semanticCorrectInfo) {
+
+
+
+    }
+}
